requirements.txt
# Main requirements for Person Detection App
# This file includes all Python dependencies for the entire project

# Core ML/AI dependencies
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
numpy>=1.21.0

# HTTP requests for model downloading
requests>=2.25.0

# Image processing
opencv-python>=4.5.0

# Data handling
pandas>=1.3.0

# Utilities
tqdm>=4.60.0
matplotlib>=3.3.0